requirements.txt
fastapi==0.111.0 
uvicorn[standard]==0.29.0 
sqlalchemy==2.0.30 
psycopg2-binary==2.9.10 
passlib[bcrypt]==1.7.4 
python-jose[cryptography]==3.3.0 
python-multipart==0.0.9 
python-multipart==0.0.9 
python-dotenv==1.0.1 
pytest==8.4.1 
httpx==0.28.1 
